Complete some of the power ups
Used a powerup abstract class that would be used as a template for all
of the other possible power ups. So far only implemented power ups that
could increase defense by some factor and increase health by some
factor. I also modified the Player class a bit to account for power ups.
For example, the power ups change the player's current stats but there
needs to be some way to keep track of all the stats without any powerups
so a attribute dictionary was included.